fix(util): SemVer dotted prereleases + ordering, TryStringSplit remainder, %.*s casts
- SemVer: parse prerelease/build BEFORE splitting the core by '.', so dotted
  identifiers survive. Previously "1.5.0-rc.7" dropped the ".7" - which is
  OpenShock's own RC tag format, so different RCs parsed equal and OTA ordering
  between them was wrong. operator< now implements semver precedence: a release
  outranks its prerelease, prerelease identifiers compare per spec (numeric <
  alphanumeric, more identifiers outrank fewer), and build metadata is ignored.
- TryStringSplit: the final slot now keeps the whole remainder instead of
  silently dropping extra fields, so IPAddressUtils rejects "1.2.3.4.5" (the
  remainder "4.5" fails Convert::ToUint8) instead of accepting it as 1.2.3.4.
- %.*s: wrap the precision argument in static_cast<int> everywhere - it is read
  as int from varargs but .length()/.size() is size_t (real UB on 64-bit).

All covered by the new util host tests.

+++ src/GatewayConnectionManager.cpp
@@ -6,11 +6,11 @@ const char* const TAG = "GatewayConnectionManager";
 
 #include "captiveportal/Manager.h"
 #include "config/Config.h"
-#include "Core.h"
 #include "GatewayClient.h"
 #include "http/JsonAPI.h"
 #include "Logging.h"
 #include "serialization/WSLocal.h"
+#include "Temporal.h"
 
 #include "SimpleMutex.h"
 
@@ -124,7 +124,7 @@ AccountLinkResultCode GatewayConnectionManager::Link(std::string_view linkCode)
 
   DestroyClient();
 
-  OS_LOGD(TAG, "Attempting to link to account using code %.*s", linkCode.length(), linkCode.data());
+  OS_LOGD(TAG, "Attempting to link to account using code %.*s", static_cast<int>(linkCode.length()), linkCode.data());
 
   if (linkCode.length() != LINK_CODE_LENGTH) {
     OS_LOGE(TAG, "Invalid link code length: expected %zu, got %zu", static_cast<size_t>(LINK_CODE_LENGTH), linkCode.length());
